updated with handlers and helpers

+++ springboot-backend/src/main/java/com/organization/record/controller/EmployeeRecordController.java
@@ -1,8 +1,9 @@
 package com.organization.record.controller;
 
 import com.organization.record.entity.EmployeeRecord;
-import com.organization.record.exception.ResourceNotFoundException;
+import com.organization.record.exception.OrganizationServiceException;
 import com.organization.record.repository.EmployeeRecordRepo;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,6 +12,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static java.util.Objects.isNull;
+
+@Slf4j
 @CrossOrigin(origins = "http://localhost:3000")
 @RestController
 @RequestMapping("employee-record")
@@ -21,12 +25,18 @@ public class EmployeeRecordController {
     // get all employee records
     @GetMapping("/get-all-employee-records")
     public List<EmployeeRecord> getAllEmployees(){
-        return employeeRecordRepo.findAllByDeletedFalse();
+        log.info("Fetching All Employee Records");
+        List<EmployeeRecord> employeeRecords = employeeRecordRepo.findAllByDeletedFalse();
+        if (isNull(employeeRecords)){
+            throw new OrganizationServiceException("Not A Single Record Found");
+        }
+        return employeeRecords;
     }
 
     // create employee record
     @PostMapping("/create-employee-record")
     public EmployeeRecord createEmployee(@RequestBody EmployeeRecord employeeRecord) {
+        log.info("creating new employee record");
         if (employeeRecord.getDeleted() == null){
         employeeRecord.setDeleted(false);
         }
@@ -36,15 +46,21 @@ public class EmployeeRecordController {
     // get employee by id
     @GetMapping("/get-employee-record/{id}")
     public ResponseEntity<EmployeeRecord> getEmployeeById(@PathVariable Long id) {
+        log.info("searching for employee with id :"+ id);
         EmployeeRecord employeeRecord = employeeRecordRepo.findByIdAndAndDeletedFalse(id);
+        if (isNull(employeeRecord)){
+            throw new OrganizationServiceException("No record found with id :"+ id);
+        }
         return ResponseEntity.ok(employeeRecord);
     }
 
     // update employee record
     @PutMapping("/update-employee-record/{id}")
     public ResponseEntity<EmployeeRecord> updateEmployee(@PathVariable Long id, @RequestBody EmployeeRecord employeeRecordDetails){
-        EmployeeRecord employeeRecord = employeeRecordRepo.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id :" + id));
+        EmployeeRecord employeeRecord = employeeRecordRepo.findByIdAndAndDeletedFalse(id);
+        if (isNull(employeeRecord)){
+            throw new OrganizationServiceException("No record found with id :"+ id);
+        }
         if (employeeRecord.getDeleted() == null){
             employeeRecord.setDeleted(false);
         }
@@ -61,9 +77,11 @@ public class EmployeeRecordController {
     // Hard delete employee record
     @DeleteMapping("/hard-delete-employee-record/{id}")
     public ResponseEntity<Map<String, Boolean>> hardDeleteEmployee(@PathVariable Long id){
-        EmployeeRecord employeeRecord = employeeRecordRepo.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id :" + id));
-
+        log.info("hard deleting employee with id :"+ id);
+        EmployeeRecord employeeRecord = employeeRecordRepo.findByIdAndAndDeletedFalse(id);
+        if (isNull(employeeRecord)){
+            throw new OrganizationServiceException("No record found with id :"+ id);
+        }
         employeeRecordRepo.delete(employeeRecord);
         Map<String, Boolean> response = new HashMap<>();
         response.put("deleted", Boolean.TRUE);
@@ -73,10 +91,11 @@ public class EmployeeRecordController {
     // Soft Delete Employee record
     @PostMapping("/soft-delete-employee-record/{id}")
     public ResponseEntity<EmployeeRecord> softDeleteEmployee(@PathVariable Long id) {
-
-        EmployeeRecord employeeRecord = employeeRecordRepo.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not exist with id :" + id));
-
+        log.info("soft deleting employee with id :"+ id);
+        EmployeeRecord employeeRecord = employeeRecordRepo.findByIdAndAndDeletedFalse(id);
+        if (isNull(employeeRecord)){
+            throw new OrganizationServiceException("No record found with id :"+ id);
+        }
         Boolean deleted = true;
         employeeRecord.setDeleted(deleted);
         employeeRecordRepo.save(employeeRecord);
